Move storage-related types to amf-based folder
